requirements.txt
fastapi
router
uvicorn
alembic
sqlalchemy
pymysql
python-dotenv
pydantic
email-validator
python-jose[cryptography]
passlib[bcrypt]

